plat/pci_bus: Change pci_device.base from u16 to unsigned long
On arm64, no ioport at all. It uses iomem instead, hence change the
base from u16 to unsigned long.

+++ plat/common/include/pci/pci_bus.h
@@ -150,7 +150,7 @@ struct pci_device {
 	struct pci_driver     *drv;
 	enum pci_device_state state;
 
-	uint16_t base;
+	unsigned long base;
 	unsigned long irq;
 };
